Move Sealed to separate file
